high voltage remove, house in home page displayed

+++ db/migrate/20141021042140_create_houses.rb
@@ -4,16 +4,16 @@ class CreateHouses < ActiveRecord::Migration
       t.references :type, index: true
       t.string :name
       t.string :address
-      t.decimal :area
-      t.decimal :mainarea
+      t.integer :area
+      t.integer :mainarea
       t.integer :piece
       t.integer :room
       t.integer :bathroom
       t.integer :toilette
-      t.decimal :price
-      t.decimal :energy
-      t.decimal :impact
-      t.decimal :tax
+      t.integer :price
+      t.integer :energy
+      t.integer :impact
+      t.integer :tax
       t.string :heating
       t.integer :construction
       t.integer :outdoorarea
